rename course-preview to onboarding
Since it's now doing non-preview nag stuff
