+++ tutor/resources/styles/components/onboarding.less
