+++ tutor/resources/styles/tutor.less
@@ -40,7 +40,7 @@
 @import './components/paging-navigation';
 @import './components/pinned-header-footer-card/sections';
 @import './components/course-calendar/index';
-@import './components/course-preview';
+@import './components/onboarding';
 @import './components/cc-dashboard/index';
 @import './components/change-student-id';
 @import './components/dialog';
