+++ tutor/src/components/onboarding/nags.js
@@ -1,9 +1,11 @@
 import previewOnlyWarning from './preview-only-warning';
 import secondSessionWarning from './second-session-warning';
 import expiredPreviewWarning from './expired-preview-warning';
+import freshlyCreatedCourse from './freshly-created-course';
 
 export default {
   previewOnlyWarning,
   secondSessionWarning,
   expiredPreviewWarning,
+  freshlyCreatedCourse,
 };
