+++ tutor/src/components/onboarding/second-session-warning.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Button } from 'react-bootstrap';
 import { observer } from 'mobx-react';
-import { PreviewWarning, NagWarning, Body, Footer } from './preview-warning';
+import { OnboardingNag, GotItOnboardingNag, NagWarning, Body, Footer } from './onboarding-nag';
 
 @observer
-export default class SecondSessionWarning extends PreviewWarning {
+export default class SecondSessionWarning extends GotItOnboardingNag {
 
   renderPrompt() {
     const { ux } = this.props;
     return (
-      <NagWarning className="second-session-prompt">
+      <OnboardingNag className="second-session-prompt">
         <Body>
           Ready to create your real course? It’s free for you and students will pay {ux.formattedStudentCost} per course per semester.
         </Body>
@@ -17,7 +17,7 @@ export default class SecondSessionWarning extends PreviewWarning {
           <Button bsStyle="primary" onClick={this.onAddCourse}>Create your course</Button>
           <Button onClick={this.onContinue}>Ask me later</Button>
         </Footer>
-      </NagWarning>
+      </OnboardingNag>
     );
   }
